Negative numbers can now be parsed, as well as Quaternions. Setting array sizes within sii files is now parsable.

+++ Sii/Parsing/NumberFormat.cs
@@ -2,7 +2,7 @@
 {
     internal enum NumberFormat
     {
-        Decimal,
+        Integer,
         Float,
         HexFloat,
     }
